% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_IMD_vs_pol.R
\name{area_IMD_vs_pol}
\alias{area_IMD_vs_pol}
\title{Plots the average IMD vs the average NOx for geographical areas (cities or
county/UA)}
\usage{
area_IMD_vs_pol(prawn_path, pollutant, area_type, year)
}
\arguments{
\item{prawn_path}{The filepath for the csv to work from}

\item{pollutant}{The pollutant name, used in graph annotations}

\item{area_type}{can be "City" or "County/UA"}

\item{year}{}
}
\description{
Plots the average IMD vs the average NOx for geographical areas (cities or
county/UA)
}
