% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_nox_histogram.R
\name{avg_nox_histogram}
\alias{avg_nox_histogram}
\title{Create a histogram of the average NOx per LSOA with the median and mean values labelled}
\usage{
avg_nox_histogram(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The pollutant used, shows in axis title, defaults to NOx}

\item{year}{The year of measurements, used in axis title, defaults to 2020}
}
\description{
Create a histogram of the average NOx per LSOA with the median and mean values labelled
}
\examples{
avg_nox_histogram()
}
\keyword{faceted,}
\keyword{sources}
