% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_processor.R
\name{bulk_processor}
\alias{bulk_processor}
\title{A function that performs everything in the package and stores the results in
a folder that is created with an appropriate procedurally generated name}
\usage{
bulk_processor(
  raster_path,
  shapefile_path,
  pollutant_data_name,
  year,
  pollutant,
  iteration = as.character(packageVersion("PRAWNS")),
  dpi = 600,
  file_format = "png",
  gotta_go_fast = "No",
  output_path = "undefined"
)
}
\arguments{
\item{raster_path}{The filepath for the raster that is to be used. This is
compatible with folders containing multiple raster layers}

\item{shapefile_path}{The filepath for the shapefile to be used}

\item{pollutant_data_name}{what the pollutant is referred to as in the raw
data, should be quoted}

\item{year}{the year the data is from, used to standardise column names,
currently doesnt work for years before 2000}

\item{pollutant}{The name of the pollutant that's being examined, this is
used in the graph names}

\item{iteration}{Used in naming the output folder, defaults to the version
of PRAWNS used but can be used to apply custom labels, should be a string}

\item{dpi}{The dpi to save graphs with, defaults to 600, the minimum for
figures in an RSC publication, higher values will take longer to process}

\item{file_format}{The file format to save in, should be all lower case and
in quotes, as it's piped straight into a ggsave function}

\item{gotta_go_fast}{Allows the code to sacrifice things that take a long
time to get it to run faster, defaults to "No"}
}
\description{
This function takes a prawns CSV and produces a summary of the geographic areas
matching an inputted parameter. Refer to the vignette setting_up_data sources
to obtain the correct input files and structure them for compatibility
}
\examples{
bulk_processor(
  raster_path="example_raster_folder",
  shapefile_path="example_shapefile_location",
  pollutant_data_name="nox",
  year=2019,
  pollutant="NOx"
  )
}
\keyword{faceted,}
\keyword{sources}
