% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_area_for_most_deprived_src.R
\name{cartesian_area_for_most_deprived_src}
\alias{cartesian_area_for_most_deprived_src}
\title{Create a graph showing how size affects emissions}
\usage{
cartesian_area_for_most_deprived_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Create a graph showing how size affects emissions
}
\examples{
size_poker()

}
\keyword{faceted,}
\keyword{sources}
