% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_deprivation_emissions_ethnicity.R
\name{cartesian_deprivation_emissions_ethnicity}
\alias{cartesian_deprivation_emissions_ethnicity}
\title{Bind ethnicity data to the prawn and create graphs based on it}
\usage{
cartesian_deprivation_emissions_ethnicity(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The pollutant being investigated, used in graph titles}

\item{year}{The year being investigated, used in graph titles}
}
\description{
Bind ethnicity data to the prawn and create graphs based on it
}
\examples{
cartesian_ethnicity_groups_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019)
}
\keyword{faceted,}
\keyword{sources}
