% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_expanse_quintile_src.R
\name{cartesian_expanse_quintile_src}
\alias{cartesian_expanse_quintile_src}
\title{Plot the relationship between IMD and emissions, grouped by LSOA size quintiles}
\usage{
cartesian_expanse_quintile_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The name of the pollutant that's being examined, this is
used in the graph names}

\item{year}{The year the dat is from, used for axis labels}
}
\description{
Plot the relationship between IMD and emissions, grouped by LSOA size quintiles
}
\examples{
expanse_probe()

}
\keyword{expanse}
