% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_expanse_src.R
\name{cartesian_expanse_src}
\alias{cartesian_expanse_src}
\title{Make a nice graph}
\usage{
cartesian_expanse_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{the filepath to the prawn csv file used for the graph}

\item{pollutant}{the pollutant looked at, used in graph titles}

\item{year}{the year of the measurement, used in axis labels}
}
\description{
Make a nice graph
}
\examples{
please no error message
}
