% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_minimalist_src.R
\name{cartesian_minimalist_src}
\alias{cartesian_minimalist_src}
\title{Make a minimalist graph suitable for a graphical abstract}
\usage{
cartesian_minimalist_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{the filepath to load data from}

\item{pollutant}{the pollutant investigated, used to name axis}

\item{year}{the year the data is from, used to name axis}
}
\description{
Make a minimalist graph suitable for a graphical abstract
}
\examples{
minimalist_graph()
}
