% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_modelled_flat.R
\name{cartesian_modelled_flat}
\alias{cartesian_modelled_flat}
\title{Plot modelled data}
\usage{
cartesian_modelled_flat(model_prawn_path, pollutant, year)
}
\arguments{
\item{model_prawn_path}{the filepath to the prawn csv file that contains only rhe target region}

\item{pollutant}{the pollutant analysed, used for naming axis}
}
\description{
Plot modelled data
}
\examples{
model_plot()
}
