% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prawns.R
\name{create_prawns}
\alias{create_prawns}
\title{A function for creating the data source for PRAWNS functions``}
\usage{
create_prawns(
  raster_path,
  shapefile_path,
  data_path,
  key_variable,
  key_variable_aliases = FALSE,
  output_path = FALSE
)
}
\arguments{
\item{raster_path}{The filepath for the raster that is to be used. This is
compatible with folders containing multiple raster layers}

\item{shapefile_path}{The filepath for the shapefile to be used}

\item{data_path}{The filepath for all csv files which contain additional data
this data should have entries that correspond to polygons in the shapefile}

\item{key_variable}{The variable that is common between the shapefile and csv
files specified in data_path}

\item{key_variable_aliases}{Any alternate ways of namingf the key variable
that are used in the different inputted tables, not necessary if the data is
already nicely formatted so defaults to FALSE}

\item{output_path}{The filepath to output to. Defaults to FALSE}
}
\description{
This function takes a raster and a shapefile, works out the average value for
each polygon in the shapefile then uses an identifier present in the
shapefile to link with other data sources. This the pollution for areas to be
worked out then linked to statistics associated with those areas.
}
\examples{
create_prawns()

}
\keyword{data}
