% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_data_flagger.R
\name{data_flagger}
\alias{data_flagger}
\title{A function showing a summary of the prawn inputted to allow the data to be roughly checked}
\usage{
data_flagger(prawn_path)
}
\arguments{
\item{prawn_path}{the location of the prawn to be analysed, should be the filepath to a csv file containing a raw prawn}
}
\description{
A function showing a summary of the prawn inputted to allow the data to be roughly checked
}
