% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decile_vs_emission_by_variable.R
\name{Decile_vs_emission_by_variable}
\alias{Decile_vs_emission_by_variable}
\title{Create a graph of emissions vs decile for a chosen variable}
\usage{
Decile_vs_emission_by_variable(
  active_stack,
  chosen_decile,
  chosen_variable,
  chosen_grouping,
  title,
  xaxis,
  yaxis,
  UK_Average = FALSE
)
}
\arguments{
\item{active_stack}{the data to use for the graph, should be an existing object in r}

\item{chosen_decile}{the decile to use for the x scale, should match a column in the data}

\item{chosen_variable}{the variable to use for the y scale, should match a column in the data}

\item{chosen_grouping}{the variable to group by, should match a column in the data}

\item{title}{the title to give th graph}

\item{xaxis}{the label for the x axis}

\item{yaxis}{the label for the y axis}

\item{UK_average}{whether to show the UK average for reference, defaults to FALSE}
}
\description{
it works off a variable already specified in R, this was chosen over a file path as
}
