% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_by_expanse.R
\name{emissions_by_expanse}
\alias{emissions_by_expanse}
\title{Make a nice graph}
\usage{
emissions_by_expanse(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{the filepath to the prawn csv file used for the graph}

\item{pollutant}{the pollutant looked at, used in graph titles}

\item{year}{the year of the measurement, used in axis labels}
}
\description{
Make a nice graph
}
