% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity_bulk_processor.R
\name{ethnicity_bulk_processor}
\alias{ethnicity_bulk_processor}
\title{Run ethnicity graphs and save results}
\usage{
ethnicity_bulk_processor(year, pollutant)
}
\description{
Run ethnicity graphs and save results
}
\examples{
cartesian_ethnicity_groups_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019)
}
