% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanse_probe.R
\name{expanse_probe}
\alias{expanse_probe}
\title{A function for looking at the relationship between IMD and emissions for
different LSOA sizes
populations}
\usage{
expanse_probe(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The name of the pollutant that's being examined, this is used in the graph names}

\item{year}{The year the dat is from, used for axis labels}
}
\description{
A function for looking at the relationship between IMD and emissions for
different LSOA sizes
populations
}
\examples{
expanse_probe()

}
\keyword{expanse}
