% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_all_areas_src.R
\name{facet_all_areas_src}
\alias{facet_all_areas_src}
\title{Creates a graph with a little window showing every item in the group separately
e.g. if grouped by city there would be a tile for every city.}
\usage{
facet_all_areas_src(prawn_path, group, pollutant, facet_name)
}
\arguments{
\item{prawn_path}{the data source to use, should be a quoted csv filepath}

\item{group}{The group to facet by, can be TCITY15NM to group by city or
anything else to group by county/UA, this is due to jerry rigging a path
around non standard evaluation}

\item{pollutant}{the name of the pollutant, used to name the graphs}

\item{facet_name}{a string used to name what the data is being faceted by in the graph title}
}
\description{
Creates a graph with a little window showing every item in the group separately
e.g. if grouped by city there would be a tile for every city.
}
\examples{
facet_all_areas_srv(
  prawn_path="PRAWN.csv",
  group="TCITY15NM",
  pollutant="NOx",
  facet_name="City")
}
