% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_ethnicity_quintile_dual_y_src.R
\name{facet_ethnicity_quintile_dual_y_src}
\alias{facet_ethnicity_quintile_dual_y_src}
\title{Bind ethnicity data to the prawn and create graphs based on it}
\usage{
facet_ethnicity_quintile_dual_y_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The pollutant being investigated, used in graph titles}

\item{year}{The year being investigated, used in graph titles}
}
\description{
Bind ethnicity data to the prawn and create graphs based on it
}
\examples{
facet_ethnicity_quintile_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019
  )
}
\keyword{faceted,}
\keyword{sources}
