% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_expanse_src.R
\name{facet_expanse_src}
\alias{facet_expanse_src}
\title{Facet into 12 size groupings of LSOA and plot emissions inequality from
each source sector}
\usage{
facet_expanse_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The name of the pollutant that's being examined, this is used in the graph names}

\item{year}{The year the dat is from, used for axis labels}
}
\description{
Facet into 12 size groupings of LSOA and plot emissions inequality from
each source sector
}
\examples{
facet_expanse_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019
  )
}
\keyword{expanse}
