% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_RUC_decile_emissions_ethnicity.R
\name{facet_RUC_decile_emissions_ethnicity}
\alias{facet_RUC_decile_emissions_ethnicity}
\title{Bind ethnicity data to the prawn and create graphs based on it}
\usage{
facet_RUC_decile_emissions_ethnicity(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The pollutant being investigated, used in graph titles}

\item{year}{The year being investigated, used in graph titles}
}
\description{
Bind ethnicity data to the prawn and create graphs based on it
}
\keyword{faceted,}
\keyword{sources}
