% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_sources_by_expanse_src.R
\name{facet_sources_by_expanse_src}
\alias{facet_sources_by_expanse_src}
\title{Plot the relationship between emissions and LSOA expanse for each emissions
sector}
\usage{
facet_sources_by_expanse_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{the filepath to the prawn csv file used for the graph}

\item{pollutant}{the pollutant looked at, used in graph titles}

\item{year}{the year of the measurement, used in axis labels}
}
\description{
Plot the relationship between emissions and LSOA expanse for each emissions
sector
}
\examples{
facet_sources_by_expanse_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019
  )
}
