% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_sources_src.R
\name{facet_sources_src}
\alias{facet_sources_src}
\title{A function for faceting the sources of a pollutant}
\usage{
facet_sources_src(
  prawn_path = "blank",
  pollutant,
  year,
  input_prawn,
  trim = TRUE
)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used, should
be quoted}

\item{pollutant}{The name of the pollutant that's being examined, this is
used in the graph names, should be a string}

\item{year}{The year of the data, used in graph labelling}

\item{input_prawn}{Use to directly input a prawns object}

\item{trim}{If trim is true then the firlds tot_area and offshore are trimmed for reduncancy}
}
\description{
This function takes a prawns CSV and produces a summary of the geographic
areas matching an inputted parameter
}
\examples{
facet_sources_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019
  )
}
\keyword{faceted,}
\keyword{sources}
