% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faceted_plot.R
\name{faceted_plot}
\alias{faceted_plot}
\title{Creates a graph wit a little window showing every item in the group separately
e.g. if grouped by city there would be a tile for every city.}
\usage{
faceted_plot(prawn_path, group, pollutant, facet_name)
}
\arguments{
\item{prawn_path}{the data source to use, should be a quoted csv filepath}

\item{group}{The group to facet by, can be TCITY15NM to group by city or
anything else to group by county/UA, this is due to jerry rigging a path
around non standard evaluation}

\item{pollutant}{the name of the pollutant, used to name the graphs}

\item{facet_name}{a string used to name what the data is being faceted by in the graph title}
}
\description{
Creates a graph wit a little window showing every item in the group separately
e.g. if grouped by city there would be a tile for every city.
}
