% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floundering.R
\name{floundering}
\alias{floundering}
\title{A function for faceting the sources of a pollutant}
\usage{
floundering(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used, should be quoted}

\item{pollutant}{The name of the pollutant that's being examined, this is used in the graph names, should be a string}

\item{year}{The year of the data, used in graph labelling}
}
\description{
This function takes a prawns CSV and produces a summary of the geographic areas
matching an inputted parameter
}
\examples{
faceted_sources()

}
\keyword{faceted,}
\keyword{sources}
