% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_area_multiplot_src.R
\name{focus_area_multiplot_src}
\alias{focus_area_multiplot_src}
\title{A function for creating a graphic summarising a city through tiling smaller
graphs}
\usage{
focus_area_multiplot_src(
  prawn_path,
  shapefile_path,
  targets,
  pollutant,
  output_path = FALSE
)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{shapefile_path}{The filepath for the shapefile that is to be used}

\item{targets}{The name of the thing you're trying to isolate e.g. "Manchester"
can also be a vector of valid values}

\item{pollutant}{The name of the pollutant that's being examined, this is used in the graph names}

\item{output_path}{The filepath to output to, a folder will be created at
this location which contains all the graphs produced by this code. Defaults
to FALSE}
}
\description{
This function takes a prawns CSV and produces a summary of the geographic areas
matching an inputted parameter
}
\keyword{graph,}
\keyword{heatmap,}
