% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focused_plot.R
\name{focused_plot}
\alias{focused_plot}
\title{Make a nice graph}
\usage{
focused_plot(focused_prawn_path, base_prawn_path, pollutant, year)
}
\arguments{
\item{focused_prawn_path}{the filepath to the prawn csv file that contains only rhe target region}

\item{base_prawn_path}{the filepath to the prawn csv with the whole data set, used to plot an average line for comparison}
}
\description{
Make a nice graph
}
\examples{
focused_plot()
}
