% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_summary.R
\name{geographic_summary}
\alias{geographic_summary}
\title{A function for creating a graphic summarising a city}
\usage{
geographic_summary(
  prawn_path,
  shape_path,
  column_targeted,
  targets,
  key_variable = "LSOA19CD",
  output_path = FALSE
)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{shape_path}{The filepath for the shapefile that is to be used}

\item{column_targeted}{The name of the column to be searched. It is reccomended
to use somehtign like city or county/UA that is geographically specific, but
any variable can be used}

\item{targets}{The name of the thing you're trying to isolate e.g. "Manchester"
can also be a vector of valid values}

\item{key_variable}{The variable that is common between the shapefile and csv
files specified in data_path, defaults to LSOA19CD}

\item{output_path}{The filepath to output to, a folder will be created at
this location which contains all the graphs produced by this code. Defaults
to FALSE}
}
\description{
This function takes a prawns CSV and produces a summary of the geographic areas
matching an inputted parameter
}
\examples{
geographic_summary()

}
\keyword{graph,}
\keyword{heatmap,}
