% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_bar.R
\name{gradient_bar}
\alias{gradient_bar}
\title{This function creates a pie chart showing how each source sector contributes
to the overall gradient as a percentage, it works off a file created by
stat_wrangler and is currently designed to work within the environment of a
bulk_processor call}
\usage{
gradient_bar(pollutant, input_path)
}
\arguments{
\item{pollutant}{the polutant being examined, used in the graph title}

\item{input_path}{the object to use as a data source defaults to FALSE this
should be the name of an existing object}
}
\description{
This function creates a pie chart showing how each source sector contributes
to the overall gradient as a percentage, it works off a file created by
stat_wrangler and is currently designed to work within the environment of a
bulk_processor call
}
\examples{
gradient_bar()
}
\keyword{data}
