% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_saver.R
\name{graph_saver}
\alias{graph_saver}
\title{Save ggplot2 graphs with sensible scaling and size options ready for publication}
\usage{
graph_saver(plot, filename, file_format, type)
}
\arguments{
\item{plot}{The ggplot2 object to plot}

\item{filename}{The location to save the file to}

\item{file_format}{The format to save using, this can include agg_png and the
agg family}

\item{type}{The graph size to save to 1 is a single column figure on a two
column page, 2 is a full page spread, both use a 16:9 aspect ratio, 3 is an
entire page}
}
\description{
Save ggplot2 graphs with sensible scaling and size options ready for publication
}
\examples{
graph_saver()
}
