% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSOA_pollutant_histo.R
\name{LSOA_pollutant_histo}
\alias{LSOA_pollutant_histo}
\title{Create a faceted histogram showing all but the top 1\% of the nox emission average across lsoas}
\usage{
LSOA_pollutant_histo(prawn_path, pollutant = "NOx", year = 2020)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Create a faceted histogram showing all but the top 1\% of the nox emission average across lsoas
}
\examples{
LSOA_pollutant_histo()
}
\keyword{faceted,}
\keyword{sources}
