% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalist_graph.R
\name{minimalist_graph}
\alias{minimalist_graph}
\title{Make a minimalist graph suitable for a graphical abstract}
\usage{
minimalist_graph(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{the filepath to load data from}

\item{pollutant}{the pollutant investigated, used to name axis}

\item{year}{the year the data is from, used to name axis}
}
\description{
Make a minimalist graph suitable for a graphical abstract
}
\examples{
minimalist_graph()
}
