% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot.R
\name{model_plot}
\alias{model_plot}
\title{Plot modelled data}
\usage{
model_plot(model_prawn_path, pollutant, year)
}
\arguments{
\item{model_prawn_path}{the filepath to the prawn csv file that contains only rhe target region}

\item{pollutant}{the pollutant analysed, used for naming axis}
}
\description{
Plot modelled data
}
\examples{
model_plot()
}
