% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NI_bulk_processor.R
\name{NI_bulk_processor}
\alias{NI_bulk_processor}
\title{A function that performs everything in the package and stores the results in
a folder that is created with an appropriate procedurally generated name}
\usage{
NI_bulk_processor(
  raster_path,
  shapefile_path,
  pollutant_data_name,
  year,
  pollutant,
  iteration = as.character(packageVersion("PRAWNS"))
)
}
\arguments{
\item{raster_path}{The filepath for the raster that is to be used. This is
compatible with folders containing multiple raster layers}

\item{shapefile_path}{The filepath for the shapefile to be used}

\item{pollutant_data_name}{what the pollutant is referred to as in the raw data, should be quoted}

\item{year}{the year the data is from, used to standardise column names, currently doesnt work for years before 2000}

\item{pollutant}{The name of the pollutant that's being examined, this is used in the graph names}

\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
This function takes a prawns CSV and produces a summary of the geographic areas
matching an inputted parameter. Refer to the vignette setting_up_data sources
to obtain the correct input files and structure them for compatibility
}
\examples{
bulk_processor()

}
\keyword{faceted,}
\keyword{sources}
