% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p values for chunks.R
\name{p_values_for_chunks}
\alias{p_values_for_chunks}
\title{Create a faceted histogram showing all but the top 1\% of the nox emission average across lsoas}
\usage{
p_values_for_chunks(prawn_path)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Create a faceted histogram showing all but the top 1\% of the nox emission average across lsoas
}
\examples{
p_values_for_chunks()
}
\keyword{faceted,}
\keyword{sources}
