% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_area_gradients.R
\name{plot_area_gradients}
\alias{plot_area_gradients}
\title{Create a histogram showign the gradients for every county/ua or city}
\usage{
plot_area_gradients(prawn_path)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Create a histogram showign the gradients for every county/ua or city
}
\examples{
plot_area_gradients()
}
\keyword{faceted,}
\keyword{sources}
