% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph_saver.R
\name{process_graph_saver}
\alias{process_graph_saver}
\title{Save ggplot2 graphs with sensible scaling and size options ready for publication}
\usage{
process_graph_saver(plot, filename, file_format, type, scaling = 0.2)
}
\arguments{
\item{plot}{The ggplot2 object to plot}

\item{filename}{The location to save the file to}

\item{file_format}{The format to save using, this can include agg_png and the
agg family}

\item{type}{The graph size to save to 1 is a single column figure on a two
column page, 2 is a full page spread, both use a 16:9 aspect ratio, 3 is an
entire page, 4 is a graphical abstract}

\item{scaling}{The =scaling parameter to use, flexible because there are many
things at play when the default is picked}
}
\description{
Save ggplot2 graphs with sensible scaling and size options ready for publication
}
