% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_shapefile_checker.R
\name{process_shapefile_checker}
\alias{process_shapefile_checker}
\title{Outputs the shapefile as a graphic so it can be visually checked}
\usage{
process_shapefile_checker(shapefile_path)
}
\arguments{
\item{shapefile_path}{the path to the shapefiles that are being checked}
}
\description{
Outputs the shapefile as a graphic so it can be visually checked
}
\examples{
process_shapefile_checker(
shapefile_path="Shapefile_location")
}
