% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tempura_prawns.R
\name{process_tempura_prawns}
\alias{process_tempura_prawns}
\title{A variant of create_prawns which uses the PRAWNSdata package to access
static data in a preprocessed form, enabling use of the tempura gui on
shinyapps`}
\usage{
process_tempura_prawns(
  raster_path = "undefined",
  tif_path = "undefined",
  csv_coordinates_path = "undefined",
  shapefile_path = "Data/2011_LSOA_shapefile_20m_generalised",
  output_path = "undefined",
  pollutant_data_name,
  year,
  pollutant,
  is_raw = FALSE
)
}
\arguments{
\item{raster_path}{The filepath for the raster that is to be used. This is
compatible with folders containing multiple raster layers, mutually exclusive with tif path}

\item{tif_path}{The filepath for the tif file used to input modelled data, mutually sxclusive with raster_path}

\item{shapefile_path}{The filepath for the shapefile to be used}

\item{output_path}{The filepath to output to. Defaults to FALSE}

\item{pollutant_data_name}{what the pollutant is referred to as in the tables}

\item{year}{the year the data is from, used to standardise column names, currently doesnt work for years before 2000}

\item{pollutant}{the pollutant being investigated, this should be in quotes and correspond with its name in the data}

\item{is_raw}{whether or not to output a csv only containing the LSOAs mapped to the emissions}
}
\description{
This function takes a raster and a shapefile, works out the average value for
each polygon in the shapefile then uses an identifier present in the
shapefile to link with other data sources. This the pollution for areas to be
worked out then linked to statistics associated with those areas. Refer to +
the vignette setting_up_data sources to obtain the correct input files and
structure them for compatibility
}
\examples{
process_tempura_prawns(
  raster_path="The folder containing the rasters",
  shapefile_path="Where the shapefiles are",
  pollutant_data_name="nox",
  year="2019",
  pollutant="NOx")
}
\keyword{data}
