% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUC_IMD_targeted.R
\name{RUC_IMD_targeted}
\alias{RUC_IMD_targeted}
\title{A function for looking at the relationship between IMD and emissions for
rural urban classifications. The output is a list with two entries output1 is
a graph of the emissions by decile for each classification, output 2 is their
populations}
\usage{
RUC_IMD_targeted(prawn_path, pollutant)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The name of the pollutant that's being examined, this is used in the graph names}
}
\description{
A function for looking at the relationship between IMD and emissions for
rural urban classifications. The output is a list with two entries output1 is
a graph of the emissions by decile for each classification, output 2 is their
populations
}
\examples{
RUC_IMD()

}
\keyword{RUC}
