% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_by_side_stats.R
\name{side_by_side_stats}
\alias{side_by_side_stats}
\title{Make a nice graph}
\usage{
side_by_side_stats(pollutant, year, prawn_path)
}
\arguments{
\item{pollutant}{the pollutant being investigated, this should be in quotes and correspond with its name in the data}

\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Make a nice graph
}
\examples{
side_by_sidestats()

}
