% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_poker.R
\name{size_poker}
\alias{size_poker}
\title{Create a graph showing how size affects emissions}
\usage{
size_poker(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Create a graph showing how size affects emissions
}
\examples{
size_poker()

}
\keyword{faceted,}
\keyword{sources}
