% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_summary.R
\name{source_summary}
\alias{source_summary}
\title{summarises the source sectors and their relationship with IMD in the data specified}
\usage{
source_summary(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The name of the pollutant that's being examined, this is used in the graph names}

\item{year}{The year used, used in graph names}
}
\description{
summarises the source sectors and their relationship with IMD in the data specified
}
