% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Startup.R
\name{Startup}
\alias{Startup}
\title{Load in all ther packages as a jerry rigged fix to figuring out namespace}
\usage{
Startup()
}
\description{
Load in all ther packages as a jerry rigged fix to figuring out namespace
}
\examples{
Startup()
}
