% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_pvalues_src.R
\name{stats_pvalues_src}
\alias{stats_pvalues_src}
\title{Create a faceted histogram showing all but the top 1\% of the nox emission average across lsoas}
\usage{
stats_pvalues_src(prawn_path)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Create a faceted histogram showing all but the top 1\% of the nox emission average across lsoas
}
\examples{
p_values_for_chunks()
}
\keyword{faceted,}
\keyword{sources}
