% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamlined_stats.R
\name{streamlined_stats}
\alias{streamlined_stats}
\title{Make a nice graph}
\usage{
streamlined_stats(pollutant, year, prawn_path = 1, input = 1)
}
\arguments{
\item{pollutant}{the pollutant being investigated, this should be in quotes and correspond with its name in the data}

\item{year}{the year the data was taken, used for axis titles}

\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Make a nice graph
}
\examples{
side_by_sidestats()

}
