% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_focus_compare_trimmed_data_src.R
\name{WIP_focus_compare_trimmed_data_src}
\alias{WIP_focus_compare_trimmed_data_src}
\title{Compare a trimmed data set with the whole, this function doesn't cover the
trimming, provide it with two prawns and it'll do the work}
\usage{
WIP_focus_compare_trimmed_data_src(
  focused_prawn_path,
  base_prawn_path,
  pollutant,
  year
)
}
\arguments{
\item{focused_prawn_path}{the filepath to the prawn csv file that contains
only the target region}

\item{base_prawn_path}{the filepath to the prawn csv with the whole data set,
used to plot an average line for comparison}

\item{pollutant}{The pollutant investigated, used to generate axis titles}

\item{year}{The year the data is from, used to generate axis titles}
}
\description{
Compare a trimmed data set with the whole, this function doesn't cover the
trimming, provide it with two prawns and it'll do the work
}
\examples{
focused_plot()
}
