% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_nox_histogram.R
\name{avg_nox_histogram}
\alias{avg_nox_histogram}
\title{Create a histogram showign the gradients for every county/ua or city}
\usage{
avg_nox_histogram(prawn_path)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Create a histogram showign the gradients for every county/ua or city
}
\examples{
avg_nox_histogram()
}
\keyword{faceted,}
\keyword{sources}
