% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_area_avg_deprivation_src.R
\name{cartesian_area_avg_deprivation_src}
\alias{cartesian_area_avg_deprivation_src}
\title{Plots the average IMD vs the average NOx for geographical areas (cities or
counties/unitary authorities)}
\usage{
cartesian_area_avg_deprivation_src(prawn_path, pollutant, area_type, year)
}
\arguments{
\item{prawn_path}{The filepath for the csv to work from}

\item{pollutant}{The pollutant name, used in graph annotations}

\item{area_type}{The type of area investigated, can be "City" or "County/UA"}

\item{year}{The year of the data, used in graph annotations}
}
\description{
Plots the average IMD vs the average NOx for geographical areas (cities or
counties/unitary authorities)
}
\examples{
  cartesian_area_avg_deprivation_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  area_type="City",
  year=2019)
}
