% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_area_gradients_src.R
\name{cartesian_area_gradients_src}
\alias{cartesian_area_gradients_src}
\title{Create a histogram showign the gradients for every county/ua or city}
\usage{
cartesian_area_gradients_src(prawn_path, area_type)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{area_type}{They type of area to group by, either city or county/UA}
}
\description{
Create a histogram showign the gradients for every county/ua or city
}
\examples{
plot_area_gradients()
}
\keyword{faceted,}
\keyword{sources}
