% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_total_emissions_src.R
\name{histogram_total_emissions_src}
\alias{histogram_total_emissions_src}
\title{Create a histogram of the average total emissions with the median and mean
values labelled}
\usage{
histogram_total_emissions_src(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The pollutant used, shows in axis title, defaults to NOx}

\item{year}{The year of measurements, used in axis title, defaults to 2020}
}
\description{
Create a histogram of the average total emissions with the median and mean
values labelled
}
\examples{
histogram_total_emissions_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019
  )
}
\keyword{faceted,}
\keyword{sources}
