% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_wrangler.R
\name{stat_wrangler}
\alias{stat_wrangler}
\title{this function returns a tibble containing the absolute and percentage difference
between the mean and median of the intercept and actual value at the top and
bottom deciles, operates from a file or an object depending on if prawn or
input path is used}
\usage{
stat_wrangler(prawn_path = FALSE, input_path = FALSE, deciles)
}
\arguments{
\item{prawn_path}{the path for the data source, defaults to FALSE this should
reference a csv file}

\item{input_path}{the object to use as a data source defaults to FALSE this
should be the name of an existing object}

\item{deciles}{the deciles targeted, should be a numeric vector of length 2 containing integers between 1 and 10}
}
\description{
this function returns a tibble containing the absolute and percentage difference
between the mean and median of the intercept and actual value at the top and
bottom deciles, operates from a file or an object depending on if prawn or
input path is used
}
\examples{
stat_wrangler()


}
\keyword{data}
