% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_pvalues_src.R
\name{stats_pvalues_src}
\alias{stats_pvalues_src}
\title{Calculate p values for a linear regression model used on each source sector.
The data is split into random chunks}
\usage{
stats_pvalues_src(prawn_path)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}
}
\description{
Calculate p values for a linear regression model used on each source sector.
The data is split into random chunks
}
\examples{
p_values_for_chunks()
}
\keyword{faceted,}
\keyword{sources}
