% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_inequalities.R
\name{temporal_inequalities}
\alias{temporal_inequalities}
\title{Plot the progression in pollutant inequality over time}
\usage{
temporal_inequalities(time_range, pollutant, name_generator, output_path = "")
}
\arguments{
\item{time_range}{the years to be used, used for tracking filenames, should
be a consecutive numerical vector}

\item{pollutant}{the pollutant being investigated, this should be in quotes
and correspond with its name in the data}

\item{name_generator}{the procedure to generate the input filename, should be
a function with year as the imput}

\item{output_path}{to specify a place to put the output, will otherwise just
output to the working directory to stay file structure agnostic}
}
\description{
Plot the progression in pollutant inequality over time
}
\examples{
temporal_inequalities()

}
